api: move extra files funcs to extra_files.go
