+++ api/extra_files.go
@@ -1,0 +1,30 @@
+package api
+
+import (
+	"github.com/labstack/echo"
+)
+
+func listExtraFiles(c echo.Context) error {
+	// TODO:
+	return nil
+}
+
+func getExtraFile(c echo.Context) error {
+	// TODO:
+	return nil
+}
+
+func addExtraFiles(c echo.Context) error {
+	// TODO:
+	return nil
+}
+
+func updateExtraFiles(c echo.Context) error {
+	// TODO:
+	return nil
+}
+
+func deleteExtraFile(c echo.Context) error {
+	// TODO:
+	return nil
+}
