+++ api/extra_files_test.go
@@ -1,0 +1,25 @@
+package api
+
+import (
+	"testing"
+)
+
+func Test_listExtraFiles(t *testing.T) {
+	t.Skip("not implemented yet")
+}
+
+func Test_getExtraFiles(t *testing.T) {
+	t.Skip("not implemented yet")
+}
+
+func Test_addExtraFiles(t *testing.T) {
+	t.Skip("not implemented yet")
+}
+
+func Test_updateExtraFiles(t *testing.T) {
+	t.Skip("not implemented yet")
+}
+
+func Test_deleteExtraFiles(t *testing.T) {
+	t.Skip("not implemented yet")
+}
